+++ doc/changelog.rst
@@ -5,9 +5,10 @@ Version 0.7.3
 -------------
 
 * convert: add ``--extract-metadata=NAME``
-* metadata: add ``mem_rss`` and ``mem_private`` metadat
 * add ``--tracemalloc`` option: use the ``tracemalloc`` module to track
   Python memory allocation and get the peak of memory usage in metadata
+* add ``--track-memory`` option: run a thread reading the memory usage
+  every millisecond and store the peak as ``mem_peak`` metadata
 
 Version 0.7.2 (2016-07-21)
 --------------------------